Refacted what was UCD\Collection into two separate classes.
It didn't make much sense to have a fluent interface where it was before; it was possible
to filter down and then resolve a character by codepoint, which is not going to be anything
close to a typical requirement here. The filtering/traversal has been split out into a
more "pure" collection class. What was the collection class have been renamed to "Database"
which is more fitting.