Redesigned & refactored the internals of FileRepository.
The existing layout resulted in thigh coupling of files to ranges. These have been split apart
and then recomposed to continue to facilitate the same ideas as before. Furthermore, the
original intent of this was to no specifically target PHP files, but keep things abstract to
the point that other file formats can be used. This refactoring has achieved this; the
PHPFileRepository is now FileRepository, and no longer references anything PHP file specific.