Avoid considering XMLRepository for use in the commands if XML is unavailable.
Re-arranged the serivce provider classes to make this simpler; we need to know the file path
to know whether or not the file exists. This requires the configuration provider to have been
registered. However the configuration provider depended on what was the service provider,
making the dependency circular. I have extracted application related services to a seperate
provider, which the configuration provider now depends on.