Adding a means to effectively snapshot the state of a Collection.
This is useful in situations where the Collection is backed by something that can
only be traversed once (i.e. a Generator). Leveraging this method results means
that it becomes possible to traverse a collection multiple times. It does,
however, come at a cost when it comes to memory consumption.