Compression of the database
The database in resources/generated is now compressed using Zlib. The
default compression of Zlib is used through the "compress.zlib://"
stream wrapper.

The PHP extension "zlib" is now required, but it allows the database to
weight 16 Mio instead of almost 2 Gio.

The overhead of the decompression is not neutral. Decompressing
"01098775-01099774!1000.php.gz" 100 times takes now 2.4s instead of
1.8s for the flat PHP file.